@@ -10,12 +10,12 @@ import com.dtstack.engine.api.vo.ScheduleJobVO;
 import com.dtstack.engine.common.exception.ErrorCode;
 import com.dtstack.engine.common.exception.RdosDefineException;
 import com.dtstack.engine.dao.ScheduleJobDao;
-import com.dtstack.engine.dao.BatchJobJobDao;
-import com.dtstack.engine.api.domain.BatchJobJob;
-import com.dtstack.engine.api.domain.BatchTaskShade;
-import com.dtstack.engine.api.dto.BatchJobJobDTO;
-import com.dtstack.engine.api.dto.BatchJobJobTaskDTO;
-import com.dtstack.engine.master.vo.BatchTaskVO;
+import com.dtstack.engine.dao.ScheduleJobJobDao;
+import com.dtstack.engine.api.domain.ScheduleJobJob;
+import com.dtstack.engine.api.domain.ScheduleTaskShade;
+import com.dtstack.engine.api.dto.ScheduleJobJobDTO;
+import com.dtstack.engine.api.dto.ScheduleJobJobTaskDTO;
+import com.dtstack.engine.master.vo.ScheduleTaskVO;
 import org.apache.commons.collections.CollectionUtils;
 import org.apache.commons.lang.StringUtils;
 import org.slf4j.Logger;
@@ -45,7 +45,7 @@ public class ScheduleJobJobService implements com.dtstack.engine.api.service.Sch
     private static final String WORKFLOW_PARENT = "0";
 
     @Autowired
-    private BatchJobJobDao batchJobJobDao;
+    private ScheduleJobJobDao scheduleJobJobDao;
 
     @Autowired
     private ScheduleJobDao scheduleJobDao;
@@ -85,8 +85,8 @@ public class ScheduleJobJobService implements com.dtstack.engine.api.service.Sch
         }
 
         // 递归获取level层的子节点
-        Map<Integer, List<BatchJobJob>> result = getSpecifiedLevelJobJobs(job.getJobKey(), level, true, null);
-        BatchJobJobDTO root = new BatchJobJobDTO();
+        Map<Integer, List<ScheduleJobJob>> result = getSpecifiedLevelJobJobs(job.getJobKey(), level, true, null);
+        ScheduleJobJobDTO root = new ScheduleJobJobDTO();
         root.setJobKey(job.getJobKey());
         getTree(root, result, 1, true);
 
@@ -95,7 +95,7 @@ public class ScheduleJobJobService implements com.dtstack.engine.api.service.Sch
         getAllJobKeys(result, allJobKeys);
 
         Map<String, ScheduleJob> keyJobMap = new HashMap<>();
-        Map<Long, BatchTaskShade> idTaskMap = new HashMap<>();
+        Map<Long, ScheduleTaskShade> idTaskMap = new HashMap<>();
         getRelationData(allJobKeys, keyJobMap, idTaskMap);
 
         //flowJobId不为0时表示是工作流中的子任务
@@ -105,7 +105,7 @@ public class ScheduleJobJobService implements com.dtstack.engine.api.service.Sch
     }
 
     private void getRelationData(Set<String> allJobKeys, Map<String, ScheduleJob> keyJobMap,
-                                 Map<Long, BatchTaskShade> idTaskMap) throws Exception {
+                                 Map<Long, ScheduleTaskShade> idTaskMap) throws Exception {
         List<Long> taskIds = new ArrayList<>();
 
         List<ScheduleJob> jobs = scheduleJobDao.listJobByJobKeys(allJobKeys);
@@ -114,24 +114,24 @@ public class ScheduleJobJobService implements com.dtstack.engine.api.service.Sch
             taskIds.add(scheduleJob.getTaskId());
         }
 
-        List<BatchTaskShade> taskShades = batchTaskShadeService.getSimpleTaskRangeAllByIds(taskIds);
+        List<ScheduleTaskShade> taskShades = batchTaskShadeService.getSimpleTaskRangeAllByIds(taskIds);
         taskShades.forEach(item -> idTaskMap.put(item.getTaskId(), item));
     }
 
-    private BatchJobJobDTO getTree(BatchJobJobDTO root, Map<Integer, List<BatchJobJob>> result, int level, boolean isChild) {
+    private ScheduleJobJobDTO getTree(ScheduleJobJobDTO root, Map<Integer, List<ScheduleJobJob>> result, int level, boolean isChild) {
         if (level <= result.size()) {
-            List<BatchJobJob> currentLevel = result.get(level);
-            List<BatchJobJobDTO> children = new ArrayList<>();
-            for (BatchJobJob jobJob : currentLevel) {
+            List<ScheduleJobJob> currentLevel = result.get(level);
+            List<ScheduleJobJobDTO> children = new ArrayList<>();
+            for (ScheduleJobJob jobJob : currentLevel) {
                 if (isChild) {
                     if (jobJob.getParentJobKey().equals(root.getJobKey())) {
-                        BatchJobJobDTO child = new BatchJobJobDTO();
+                        ScheduleJobJobDTO child = new ScheduleJobJobDTO();
                         child.setJobKey(jobJob.getJobKey());
                         children.add(child);
                     }
                 } else {
                     if (jobJob.getJobKey().equals(root.getJobKey())) {
-                        BatchJobJobDTO parent = new BatchJobJobDTO();
+                        ScheduleJobJobDTO parent = new ScheduleJobJobDTO();
                         parent.setJobKey(jobJob.getParentJobKey());
                         children.add(parent);
                     }
@@ -139,7 +139,7 @@ public class ScheduleJobJobService implements com.dtstack.engine.api.service.Sch
             }
 
             int nextLevel = ++level;
-            for (BatchJobJobDTO child : children) {
+            for (ScheduleJobJobDTO child : children) {
                 getTree(child, result, nextLevel, isChild);
             }
 
@@ -156,10 +156,10 @@ public class ScheduleJobJobService implements com.dtstack.engine.api.service.Sch
      *                            在业务上不需要展示从工作流子节点到父节点的关系
      * @return
      */
-    private Map<Integer, List<BatchJobJob>> getSpecifiedLevelJobJobs(String rootKey, int level, boolean getChild, String parentFlowJobJobKey) {
+    private Map<Integer, List<ScheduleJobJob>> getSpecifiedLevelJobJobs(String rootKey, int level, boolean getChild, String parentFlowJobJobKey) {
         ScheduleJob rootJob = scheduleJobDao.getByJobKey(rootKey);
 
-        Map<Integer, List<BatchJobJob>> result = new HashMap<>();
+        Map<Integer, List<ScheduleJobJob>> result = new HashMap<>();
         List<String> jobKeys = new ArrayList<>();
         List<Long> taskIdList = new ArrayList<>();
 
@@ -167,11 +167,11 @@ public class ScheduleJobJobService implements com.dtstack.engine.api.service.Sch
         taskIdList.add(rootJob.getTaskId());
         int j = 1;
         for (int i = level; i > 0; i--) {
-            List<BatchJobJobTaskDTO> jobJobs;
+            List<ScheduleJobJobTaskDTO> jobJobs;
             if (getChild) {
-                jobJobs = batchJobJobDao.listByParentJobKeysWithOutSelfTask(jobKeys);
+                jobJobs = scheduleJobJobDao.listByParentJobKeysWithOutSelfTask(jobKeys);
             } else {
-                jobJobs = batchJobJobDao.listByJobKeysWithOutSelfTask(jobKeys, taskIdList);
+                jobJobs = scheduleJobJobDao.listByJobKeysWithOutSelfTask(jobKeys, taskIdList);
             }
 
             jobJobs = jobJobs.stream().filter(jobjob -> {
@@ -189,7 +189,7 @@ public class ScheduleJobJobService implements com.dtstack.engine.api.service.Sch
             }
 
             jobKeys = new ArrayList<>();
-            for (BatchJobJobTaskDTO jobJob : jobJobs) {
+            for (ScheduleJobJobTaskDTO jobJob : jobJobs) {
                 if (getChild) {
                     jobKeys.add(jobJob.getJobKey());
                     taskIdList.add(jobJob.getTaskId());
@@ -199,7 +199,7 @@ public class ScheduleJobJobService implements com.dtstack.engine.api.service.Sch
                 }
             }
 
-            List<BatchJobJob> jobJobList = jobJobs.stream().map(BatchJobJobTaskDTO::toJobJob).collect(Collectors.toList());
+            List<ScheduleJobJob> jobJobList = jobJobs.stream().map(ScheduleJobJobTaskDTO::toJobJob).collect(Collectors.toList());
             result.put(j, jobJobList);
             j++;
         }
@@ -219,18 +219,18 @@ public class ScheduleJobJobService implements com.dtstack.engine.api.service.Sch
     private com.dtstack.engine.master.vo.ScheduleJobVO displayOffSpringForFlowWork(ScheduleJob flowJob) throws Exception {
         com.dtstack.engine.master.vo.ScheduleJobVO vo = null;
         // 递归获取level层的子节点
-        Map<Integer, List<BatchJobJob>> result = getSpecifiedLevelJobJobs(flowJob.getJobKey(), Integer.MAX_VALUE, true, null);
-        List<BatchJobJob> firstLevel = result.get(1);
+        Map<Integer, List<ScheduleJobJob>> result = getSpecifiedLevelJobJobs(flowJob.getJobKey(), Integer.MAX_VALUE, true, null);
+        List<ScheduleJobJob> firstLevel = result.get(1);
         if (CollectionUtils.isNotEmpty(firstLevel)) {
             Set<String> allJobKeys = new HashSet<>();
             getAllJobKeys(result, allJobKeys);
 
             Map<String, ScheduleJob> keyJobMap = new HashMap<>();
-            Map<Long, BatchTaskShade> idTaskMap = new HashMap<>();
+            Map<Long, ScheduleTaskShade> idTaskMap = new HashMap<>();
             getRelationData(allJobKeys, keyJobMap, idTaskMap);
 
-            BatchJobJob beginJobJob = null;
-            for (BatchJobJob jobjob : firstLevel) {
+            ScheduleJobJob beginJobJob = null;
+            for (ScheduleJobJob jobjob : firstLevel) {
                 ScheduleJob subJob = keyJobMap.get(jobjob.getJobKey());
                 //找出工作流的起始节点
                 if (StringUtils.equals(subJob.getFlowJobId(), flowJob.getJobId())) {
@@ -244,7 +244,7 @@ public class ScheduleJobJobService implements com.dtstack.engine.api.service.Sch
                 return null;
             }
 
-            BatchJobJobDTO root = new BatchJobJobDTO();
+            ScheduleJobJobDTO root = new ScheduleJobJobDTO();
             root.setJobKey(beginJobJob.getJobKey());
             //第一层作为root
             getTree(root, result, 2, true);
@@ -256,11 +256,11 @@ public class ScheduleJobJobService implements com.dtstack.engine.api.service.Sch
     }
 
     @Forbidden
-    public com.dtstack.engine.master.vo.ScheduleJobVO getOffSpring(BatchJobJobDTO root, Map<String, ScheduleJob> keyJobMap, Map<Long, BatchTaskShade> idTaskMap, boolean isSubTask) {
+    public com.dtstack.engine.master.vo.ScheduleJobVO getOffSpring(ScheduleJobJobDTO root, Map<String, ScheduleJob> keyJobMap, Map<Long, ScheduleTaskShade> idTaskMap, boolean isSubTask) {
         ScheduleJob job = keyJobMap.get(root.getJobKey());
         com.dtstack.engine.master.vo.ScheduleJobVO vo = new com.dtstack.engine.master.vo.ScheduleJobVO(job);
         vo.setProjectId(job.getProjectId());
-        BatchTaskShade batchTaskShade = idTaskMap.get(job.getTaskId());
+        ScheduleTaskShade batchTaskShade = idTaskMap.get(job.getTaskId());
         if (batchTaskShade == null) {
             return null;
         }
@@ -274,9 +274,9 @@ public class ScheduleJobJobService implements com.dtstack.engine.api.service.Sch
 
         vo.setBatchTask(getTaskVo(batchTaskShade, job));
         if (CollectionUtils.isNotEmpty(root.getChildren())) {
-            Iterator<BatchJobJobDTO> it = root.getChildren().iterator();
+            Iterator<ScheduleJobJobDTO> it = root.getChildren().iterator();
             while (it.hasNext()) {
-                BatchJobJobDTO jobJob = it.next();
+                ScheduleJobJobDTO jobJob = it.next();
 
                 if (job.getTaskId().longValue() == batchJobService.getTaskIdFromJobKey(jobJob.getJobKey()).longValue()) {
                     it.remove();
@@ -291,7 +291,7 @@ public class ScheduleJobJobService implements com.dtstack.engine.api.service.Sch
             }
 
             List<ScheduleJobVO> subJobVOs = new ArrayList<>(root.getChildren().size());
-            for (BatchJobJobDTO jobJobDTO : root.getChildren()) {
+            for (ScheduleJobJobDTO jobJobDTO : root.getChildren()) {
                 com.dtstack.engine.master.vo.ScheduleJobVO subVO = getOffSpring(jobJobDTO, keyJobMap, idTaskMap, isSubTask);
                 if (subVO != null) {
                     subJobVOs.add(subVO);
@@ -304,8 +304,8 @@ public class ScheduleJobJobService implements com.dtstack.engine.api.service.Sch
     }
 
 
-    private BatchTaskVO getTaskVo(BatchTaskShade batchTaskShade, ScheduleJob job) {
-        BatchTaskVO taskVO = new BatchTaskVO(batchTaskShade, true);
+    private ScheduleTaskVO getTaskVo(ScheduleTaskShade batchTaskShade, ScheduleJob job) {
+        ScheduleTaskVO taskVO = new ScheduleTaskVO(batchTaskShade, true);
         return taskVO;
     }
 
@@ -314,9 +314,9 @@ public class ScheduleJobJobService implements com.dtstack.engine.api.service.Sch
      */
     public com.dtstack.engine.master.vo.ScheduleJobVO displayOffSpringWorkFlow(@Param("jobId") Long jobId, @Param("appType")Integer appType) throws Exception {
         ScheduleJob job = batchJobService.getJobById(jobId);
-        BatchTaskShade batchTaskShade = batchTaskShadeService.getBatchTaskById(job.getTaskId(),appType);
+        ScheduleTaskShade batchTaskShade = batchTaskShadeService.getBatchTaskById(job.getTaskId(),appType);
         com.dtstack.engine.master.vo.ScheduleJobVO vo = new com.dtstack.engine.master.vo.ScheduleJobVO(job);
-        vo.setBatchTask(new BatchTaskVO(batchTaskShade, true));
+        vo.setBatchTask(new ScheduleTaskVO(batchTaskShade, true));
         if (batchTaskShade.getTaskType().intValue() == EJobType.WORK_FLOW.getVal() || batchTaskShade.getTaskType().intValue() == EJobType.ALGORITHM_LAB.getVal()) {
             try {
                 //工作流下全部实例,层级level使用int最大值
@@ -329,9 +329,9 @@ public class ScheduleJobJobService implements com.dtstack.engine.api.service.Sch
         return vo;
     }
 
-    private void getAllJobKeys(Map<Integer, List<BatchJobJob>> result, Set<String> jobKeys) {
+    private void getAllJobKeys(Map<Integer, List<ScheduleJobJob>> result, Set<String> jobKeys) {
         result.forEach((key, value) -> {
-            for (BatchJobJob jobJob : value) {
+            for (ScheduleJobJob jobJob : value) {
                 jobKeys.add(jobJob.getJobKey());
                 jobKeys.add(jobJob.getParentJobKey());
             }
@@ -356,8 +356,8 @@ public class ScheduleJobJobService implements com.dtstack.engine.api.service.Sch
         }
 
         // 递归获取level层的子节点
-        Map<Integer, List<BatchJobJob>> result = getSpecifiedLevelJobJobs(job.getJobKey(), level, false, parentFlowJobJobKey);
-        BatchJobJobDTO root = new BatchJobJobDTO();
+        Map<Integer, List<ScheduleJobJob>> result = getSpecifiedLevelJobJobs(job.getJobKey(), level, false, parentFlowJobJobKey);
+        ScheduleJobJobDTO root = new ScheduleJobJobDTO();
         root.setJobKey(job.getJobKey());
         getTree(root, result, 1, false);
 
@@ -366,20 +366,20 @@ public class ScheduleJobJobService implements com.dtstack.engine.api.service.Sch
         getAllJobKeys(result, allJobKeys);
 
         Map<String, ScheduleJob> keyJobMap = new HashMap<>();
-        Map<Long, BatchTaskShade> idTaskMap = new HashMap<>();
+        Map<Long, ScheduleTaskShade> idTaskMap = new HashMap<>();
         getRelationData(allJobKeys, keyJobMap, idTaskMap);
 
         return getForefathers(root, keyJobMap, idTaskMap);
     }
 
-    private com.dtstack.engine.master.vo.ScheduleJobVO getForefathers(BatchJobJobDTO root, Map<String, ScheduleJob> keyJobMap,
-                                                                      Map<Long, BatchTaskShade> idTaskMap) {
+    private com.dtstack.engine.master.vo.ScheduleJobVO getForefathers(ScheduleJobJobDTO root, Map<String, ScheduleJob> keyJobMap,
+                                                                      Map<Long, ScheduleTaskShade> idTaskMap) {
         ScheduleJob job = keyJobMap.get(root.getJobKey());
         if (job == null) {
             return null;
         }
         com.dtstack.engine.master.vo.ScheduleJobVO vo = new com.dtstack.engine.master.vo.ScheduleJobVO(job);
-        BatchTaskShade batchTaskShade = idTaskMap.get(job.getTaskId());
+        ScheduleTaskShade batchTaskShade = idTaskMap.get(job.getTaskId());
 
         vo.setBatchTask(getTaskVo(batchTaskShade, job));
 
@@ -391,7 +391,7 @@ public class ScheduleJobJobService implements com.dtstack.engine.api.service.Sch
             root.getChildren().removeIf(jobJobDTO -> job.getTaskId().equals(batchJobService.getTaskIdFromJobKey(jobJobDTO.getJobKey())));
 
             List<ScheduleJobVO> fatherVOs = new ArrayList<>();
-            for (BatchJobJobDTO jobJobDTO : root.getChildren()) {
+            for (ScheduleJobJobDTO jobJobDTO : root.getChildren()) {
                 com.dtstack.engine.master.vo.ScheduleJobVO item = this.getForefathers(jobJobDTO, keyJobMap, idTaskMap);
                 if (item != null) {
                     fatherVOs.add(item);
@@ -404,12 +404,12 @@ public class ScheduleJobJobService implements com.dtstack.engine.api.service.Sch
     }
 
     @Forbidden
-    public List<BatchJobJob> getJobChild(String parentJobKey) {
-        return batchJobJobDao.listByParentJobKey(parentJobKey);
+    public List<ScheduleJobJob> getJobChild(String parentJobKey) {
+        return scheduleJobJobDao.listByParentJobKey(parentJobKey);
     }
 
-    public int batchInsert(List<BatchJobJob> batchJobJobs) {
-        return batchJobJobDao.batchInsert(batchJobJobs);
+    public int batchInsert(List<ScheduleJobJob> scheduleJobJobs) {
+        return scheduleJobJobDao.batchInsert(scheduleJobJobs);
     }
 
 }
