@@ -1,6 +1,6 @@
 package com.dtstack.engine.dao;
 
-import com.dtstack.engine.api.domain.StreamTaskCheckpoint;
+import com.dtstack.engine.api.domain.EngineJobCheckpoint;
 import org.apache.ibatis.annotations.Param;
 
 import java.sql.Timestamp;
@@ -11,7 +11,7 @@ import java.util.List;
  * author: toutian
  * create: 2020/02/12
  */
-public interface StreamTaskCheckpointDao {
+public interface EngineJobCheckpointDao {
 
     int insert(@Param("taskId")String taskId, @Param("engineTaskId")String engineTaskId,
                @Param("checkpointId") String checkpointId,
@@ -19,23 +19,23 @@ public interface StreamTaskCheckpointDao {
                @Param("checkpointSavepath") String checkpointSavepath,
                @Param("checkpointCounts") String checkpointCounts);
 
-    List<StreamTaskCheckpoint> listByTaskIdAndRangeTime(@Param("taskId") String taskEngineId,
-                                                        @Param("triggerStart") Long triggerStart,
-                                                        @Param("triggerEnd") Long triggerEnd);
+    List<EngineJobCheckpoint> listByTaskIdAndRangeTime(@Param("taskId") String taskEngineId,
+                                                       @Param("triggerStart") Long triggerStart,
+                                                       @Param("triggerEnd") Long triggerEnd);
 
-    StreamTaskCheckpoint getByTaskIdAndEngineTaskId(@Param("taskId") String taskId, @Param("taskEngineId") String taskEngineId);
+    EngineJobCheckpoint getByTaskIdAndEngineTaskId(@Param("taskId") String taskId, @Param("taskEngineId") String taskEngineId);
 
     void batchDeleteByEngineTaskIdAndCheckpointId(@Param("taskEngineId") String taskEngineId, @Param("checkpointId") String checkpointId);
 
-    List<StreamTaskCheckpoint> getByTaskEngineIdAndCheckpointIndexAndCount(@Param("taskEngineId") String taskEngineId,
-                                                                           @Param("startIndex") int startIndex,
-                                                                           @Param("count") int count);
+    List<EngineJobCheckpoint> getByTaskEngineIdAndCheckpointIndexAndCount(@Param("taskEngineId") String taskEngineId,
+                                                                          @Param("startIndex") int startIndex,
+                                                                          @Param("count") int count);
 
     void cleanAllCheckpointByTaskEngineId(@Param("taskEngineId")  String taskEngineId);
 
     Integer updateCheckpoint(@Param("taskId") String taskId, @Param("checkpoint") String checkpoint);
 
-    StreamTaskCheckpoint getByTaskId(@Param("taskId") String taskId);
+    EngineJobCheckpoint getByTaskId(@Param("taskId") String taskId);
 
     Integer deleteByTaskId(@Param("taskId") String taskId);
 }
