@@ -1,7 +1,7 @@
 package com.dtstack.engine.master.bo;
 
 import com.dtstack.engine.api.domain.ScheduleJob;
-import com.dtstack.engine.api.domain.BatchJobJob;
+import com.dtstack.engine.api.domain.ScheduleJobJob;
 import com.google.common.collect.Lists;
 
 import java.util.List;
@@ -18,7 +18,7 @@ public class ScheduleBatchJob {
 
     private ScheduleJob scheduleJob;
 
-    private List<BatchJobJob> jobJobList = Lists.newArrayList();
+    private List<ScheduleJobJob> jobJobList = Lists.newArrayList();
 
     /**
      * 下游的任务上一个周期依赖-->null表示还未从数据库拉取过数据
@@ -33,16 +33,16 @@ public class ScheduleBatchJob {
         return scheduleJob;
     }
 
-    public List<BatchJobJob> getBatchJobJobList() {
+    public List<ScheduleJobJob> getBatchJobJobList() {
         return jobJobList;
     }
 
-    public void setJobJobList(List<BatchJobJob> jobJobList) {
+    public void setJobJobList(List<ScheduleJobJob> jobJobList) {
         this.jobJobList = jobJobList;
     }
 
-    public void addBatchJobJob(BatchJobJob batchJobJob) {
-        jobJobList.add(batchJobJob);
+    public void addBatchJobJob(ScheduleJobJob scheduleJobJob) {
+        jobJobList.add(scheduleJobJob);
     }
 
     public Long getId() {
