@@ -4,13 +4,13 @@ import com.alibaba.fastjson.JSONObject;
 import com.dtstack.dtcenter.common.annotation.Forbidden;
 import com.dtstack.dtcenter.common.enums.EJobType;
 import com.dtstack.engine.api.annotation.Param;
+import com.dtstack.engine.api.domain.ScheduleTaskTaskShade;
+import com.dtstack.engine.api.vo.ScheduleTaskVO;
 import com.dtstack.engine.common.enums.DisplayDirect;
 import com.dtstack.engine.common.exception.ErrorCode;
 import com.dtstack.engine.common.exception.RdosDefineException;
-import com.dtstack.engine.dao.BatchTaskTaskShadeDao;
-import com.dtstack.engine.api.domain.BatchTaskShade;
-import com.dtstack.engine.api.domain.BatchTaskTaskShade;
-import com.dtstack.engine.master.vo.BatchTaskVO;
+import com.dtstack.engine.dao.ScheduleTaskTaskShadeDao;
+import com.dtstack.engine.api.domain.ScheduleTaskShade;
 import com.google.common.base.Preconditions;
 import org.apache.commons.collections.CollectionUtils;
 import org.apache.commons.lang3.StringUtils;
@@ -31,13 +31,13 @@ public class ScheduleTaskTaskShadeService implements com.dtstack.engine.api.serv
     private static final Long IS_WORK_FLOW_SUBNODE = 0L;
 
     @Autowired
-    private BatchTaskTaskShadeDao batchTaskTaskShadeDao;
+    private ScheduleTaskTaskShadeDao scheduleTaskTaskShadeDao;
 
     @Autowired
     private ScheduleTaskShadeService taskShadeService;
 
     public void clearDataByTaskId(@Param("taskId") Long taskId,@Param("appType")Integer appType) {
-        batchTaskTaskShadeDao.deleteByTaskId(taskId,appType);
+        scheduleTaskTaskShadeDao.deleteByTaskId(taskId,appType);
     }
 
     @Transactional
@@ -45,35 +45,35 @@ public class ScheduleTaskTaskShadeService implements com.dtstack.engine.api.serv
         if(StringUtils.isBlank(taskLists)){
             return;
         }
-        List<BatchTaskTaskShade> taskTaskList = JSONObject.parseArray(taskLists, BatchTaskTaskShade.class);
-        Map<String,BatchTaskTaskShade> keys = new HashMap<>();
+        List<ScheduleTaskTaskShade> taskTaskList = JSONObject.parseArray(taskLists, ScheduleTaskTaskShade.class);
+        Map<String, ScheduleTaskTaskShade> keys = new HashMap<>();
         // 去重
-        for (BatchTaskTaskShade batchTaskTaskShade : taskTaskList) {
-            keys.put(String.format("%s.%s.%s",batchTaskTaskShade.getTaskId(),batchTaskTaskShade.getParentTaskId(),batchTaskTaskShade.getProjectId()),batchTaskTaskShade);
-            Preconditions.checkNotNull(batchTaskTaskShade.getTaskId());
-            Preconditions.checkNotNull(batchTaskTaskShade.getAppType());
+        for (ScheduleTaskTaskShade scheduleTaskTaskShade : taskTaskList) {
+            keys.put(String.format("%s.%s.%s", scheduleTaskTaskShade.getTaskId(), scheduleTaskTaskShade.getParentTaskId(), scheduleTaskTaskShade.getProjectId()), scheduleTaskTaskShade);
+            Preconditions.checkNotNull(scheduleTaskTaskShade.getTaskId());
+            Preconditions.checkNotNull(scheduleTaskTaskShade.getAppType());
             //清除原来关系
-            batchTaskTaskShadeDao.deleteByTaskId(batchTaskTaskShade.getTaskId(), batchTaskTaskShade.getAppType());
+            scheduleTaskTaskShadeDao.deleteByTaskId(scheduleTaskTaskShade.getTaskId(), scheduleTaskTaskShade.getAppType());
         }
 
         // 保存现有任务关系
-        for (BatchTaskTaskShade taskTaskShade : keys.values()) {
-            batchTaskTaskShadeDao.insert(taskTaskShade);
+        for (ScheduleTaskTaskShade taskTaskShade : keys.values()) {
+            scheduleTaskTaskShadeDao.insert(taskTaskShade);
         }
     }
 
-    public List<BatchTaskTaskShade> getAllParentTask(@Param("taskId") Long taskId) {
-        return batchTaskTaskShadeDao.listParentTask(taskId);
+    public List<ScheduleTaskTaskShade> getAllParentTask(@Param("taskId") Long taskId) {
+        return scheduleTaskTaskShadeDao.listParentTask(taskId);
     }
 
 
-    public BatchTaskVO displayOffSpring(@Param("taskId") Long taskId,
-                                        @Param("projectId") Long projectId,
-                                        @Param("userId") Long userId,
-                                        @Param("level") Integer level,
-                                        @Param("type") Integer directType,@Param("appType")Integer appType) {
+    public com.dtstack.engine.master.vo.ScheduleTaskVO displayOffSpring(@Param("taskId") Long taskId,
+                                                                        @Param("projectId") Long projectId,
+                                                                        @Param("userId") Long userId,
+                                                                        @Param("level") Integer level,
+                                                                        @Param("type") Integer directType, @Param("appType")Integer appType) {
 
-        BatchTaskShade task = null;
+        ScheduleTaskShade task = null;
         try {
             task = taskShadeService.getBatchTaskById(taskId,appType);
         } catch (RdosDefineException rdosDefineException) {
@@ -102,12 +102,12 @@ public class ScheduleTaskTaskShadeService implements com.dtstack.engine.api.serv
      * 0 展开上下游, 1:展开上游 2:展开下游
      * @author toutian
      */
-    private BatchTaskVO getOffSpring(BatchTaskShade taskShade, int level, Integer directType, Long currentProjectId,Integer appType) {
+    private com.dtstack.engine.master.vo.ScheduleTaskVO getOffSpring(ScheduleTaskShade taskShade, int level, Integer directType, Long currentProjectId, Integer appType) {
 
-        BatchTaskVO vo = new BatchTaskVO(taskShade, true);
+        com.dtstack.engine.master.vo.ScheduleTaskVO vo = new com.dtstack.engine.master.vo.ScheduleTaskVO(taskShade, true);
         vo.setCurrentProject(currentProjectId.equals(taskShade.getProjectId()));
         if (taskShade.getTaskType().intValue() == EJobType.WORK_FLOW.getVal()) {
-            BatchTaskVO subTaskVO = getAllFlowSubTasks(taskShade.getTaskId(),taskShade.getAppType());
+            com.dtstack.engine.master.vo.ScheduleTaskVO subTaskVO = getAllFlowSubTasks(taskShade.getTaskId(),taskShade.getAppType());
             vo.setSubNodes(subTaskVO);
         }
         if (level == 0) {
@@ -116,8 +116,8 @@ public class ScheduleTaskTaskShadeService implements com.dtstack.engine.api.serv
 
         level--;
 
-        List<BatchTaskTaskShade> taskTasks = null;
-        List<BatchTaskTaskShade> childTaskTasks = null;
+        List<ScheduleTaskTaskShade> taskTasks = null;
+        List<ScheduleTaskTaskShade> childTaskTasks = null;
 
         if(taskShade.getTaskType().intValue() != EJobType.WORK_FLOW.getVal() &&
                 !taskShade.getFlowId().equals(IS_WORK_FLOW_SUBNODE)){
@@ -126,19 +126,19 @@ public class ScheduleTaskTaskShadeService implements com.dtstack.engine.api.serv
         }
 
         if(DisplayDirect.FATHER_CHILD.getType().equals(directType) || DisplayDirect.FATHER.getType().equals(directType)){//展开上游节点
-            taskTasks = batchTaskTaskShadeDao.listParentTask(taskShade.getTaskId());
+            taskTasks = scheduleTaskTaskShadeDao.listParentTask(taskShade.getTaskId());
         }
 
         if(DisplayDirect.FATHER_CHILD.getType().equals(directType) || DisplayDirect.CHILD.getType().equals(directType)){//展开下游节点
-            childTaskTasks = batchTaskTaskShadeDao.listChildTask(taskShade.getTaskId());
+            childTaskTasks = scheduleTaskTaskShadeDao.listChildTask(taskShade.getTaskId());
         }
 
         if (CollectionUtils.isEmpty(taskTasks) && CollectionUtils.isEmpty(childTaskTasks)) {
             return vo;
         }
 
-        List<com.dtstack.engine.api.vo.BatchTaskVO> parentTaskList = null;
-        List<com.dtstack.engine.api.vo.BatchTaskVO> childTaskList = null;
+        List<ScheduleTaskVO> parentTaskList = null;
+        List<ScheduleTaskVO> childTaskList = null;
         if(!CollectionUtils.isEmpty(taskTasks)){
             Set<Long> taskIds = new HashSet<>(taskTasks.size());
             taskTasks.forEach(taskTask -> taskIds.add(taskTask.getParentTaskId()));
@@ -160,16 +160,16 @@ public class ScheduleTaskTaskShadeService implements com.dtstack.engine.api.serv
         return vo;
     }
 
-    public List<com.dtstack.engine.api.vo.BatchTaskVO> getRefTask(Set<Long> taskIds, int level, Integer directType, Long currentProjectId, Integer appType){
+    public List<ScheduleTaskVO> getRefTask(Set<Long> taskIds, int level, Integer directType, Long currentProjectId, Integer appType){
 
         //获得所有父节点task
-        List<BatchTaskShade> tasks = taskShadeService.getTaskByIds(new ArrayList<>(taskIds),appType);
+        List<ScheduleTaskShade> tasks = taskShadeService.getTaskByIds(new ArrayList<>(taskIds),appType);
         if (CollectionUtils.isEmpty(tasks)) {
             return null;
         }
 
-        List<com.dtstack.engine.api.vo.BatchTaskVO> refTaskVoList = new ArrayList<>(tasks.size());
-        for (BatchTaskShade task : tasks) {
+        List<ScheduleTaskVO> refTaskVoList = new ArrayList<>(tasks.size());
+        for (ScheduleTaskShade task : tasks) {
             refTaskVoList.add(this.getOffSpring(task, level, directType, currentProjectId,appType));
         }
 
@@ -184,9 +184,9 @@ public class ScheduleTaskTaskShadeService implements com.dtstack.engine.api.serv
      * @return
      */
     @Forbidden
-    private BatchTaskVO getOnlyAllFlowSubTasks(Long flowId,Integer appType) {
-        BatchTaskVO vo = new BatchTaskVO();
-        BatchTaskShade beginTaskShade = taskShadeService.getWorkFlowTopNode(flowId);
+    private com.dtstack.engine.master.vo.ScheduleTaskVO getOnlyAllFlowSubTasks(Long flowId, Integer appType) {
+        com.dtstack.engine.master.vo.ScheduleTaskVO vo = new com.dtstack.engine.master.vo.ScheduleTaskVO();
+        ScheduleTaskShade beginTaskShade = taskShadeService.getWorkFlowTopNode(flowId);
         if(beginTaskShade!=null) {
             vo = getFlowWorkOffSpring(beginTaskShade, 1, DisplayDirect.CHILD.getType(),appType);
         }
@@ -200,15 +200,15 @@ public class ScheduleTaskTaskShadeService implements com.dtstack.engine.api.serv
      * @param taskId
      * @return
      */
-    public BatchTaskVO getAllFlowSubTasks(@Param("taskId") Long taskId,@Param("appType") Integer appType) {
-        BatchTaskShade task = taskShadeService.getBatchTaskById(taskId,appType);
+    public com.dtstack.engine.master.vo.ScheduleTaskVO getAllFlowSubTasks(@Param("taskId") Long taskId, @Param("appType") Integer appType) {
+        ScheduleTaskShade task = taskShadeService.getBatchTaskById(taskId,appType);
         if (task == null) {
             return null;
         }
-        BatchTaskVO parentNode = new BatchTaskVO(task, true);
-        BatchTaskVO vo = new BatchTaskVO();
+        com.dtstack.engine.master.vo.ScheduleTaskVO parentNode = new com.dtstack.engine.master.vo.ScheduleTaskVO(task, true);
+        com.dtstack.engine.master.vo.ScheduleTaskVO vo = new com.dtstack.engine.master.vo.ScheduleTaskVO();
 
-        BatchTaskShade beginTaskShade = taskShadeService.getWorkFlowTopNode(taskId);
+        ScheduleTaskShade beginTaskShade = taskShadeService.getWorkFlowTopNode(taskId);
         if(beginTaskShade!=null) {
             vo = getFlowWorkOffSpring(beginTaskShade, 1, DisplayDirect.CHILD.getType(),appType);
         }
@@ -224,16 +224,16 @@ public class ScheduleTaskTaskShadeService implements com.dtstack.engine.api.serv
      * @return
      */
     @Forbidden
-    public BatchTaskVO getFlowWorkOffSpring(BatchTaskShade taskShade, int level, Integer directType,Integer appType) {
-        BatchTaskVO vo = new BatchTaskVO(taskShade, true);
-        List<BatchTaskTaskShade> childTaskTasks = null;
-        childTaskTasks = batchTaskTaskShadeDao.listChildTask(taskShade.getTaskId());
+    public com.dtstack.engine.master.vo.ScheduleTaskVO getFlowWorkOffSpring(ScheduleTaskShade taskShade, int level, Integer directType, Integer appType) {
+        com.dtstack.engine.master.vo.ScheduleTaskVO vo = new com.dtstack.engine.master.vo.ScheduleTaskVO(taskShade, true);
+        List<ScheduleTaskTaskShade> childTaskTasks = null;
+        childTaskTasks = scheduleTaskTaskShadeDao.listChildTask(taskShade.getTaskId());
         if (CollectionUtils.isEmpty(childTaskTasks)) {
             return vo;
         }
         Set<Long> taskIds = new HashSet<>(childTaskTasks.size());
         childTaskTasks.forEach(taskTask -> taskIds.add(taskTask.getTaskId()));
-        List<com.dtstack.engine.api.vo.BatchTaskVO> childTaskList = getFlowWorkSubTasksRefTask(taskIds, level, DisplayDirect.CHILD.getType(),appType);
+        List<ScheduleTaskVO> childTaskList = getFlowWorkSubTasksRefTask(taskIds, level, DisplayDirect.CHILD.getType(),appType);
         if (childTaskList != null) {
             vo.setSubTaskVOS(childTaskList);
         }
@@ -241,16 +241,16 @@ public class ScheduleTaskTaskShadeService implements com.dtstack.engine.api.serv
     }
 
     @Forbidden
-    public List<com.dtstack.engine.api.vo.BatchTaskVO> getFlowWorkSubTasksRefTask(Set<Long> taskIds, int level, Integer directType, Integer appType) {
+    public List<ScheduleTaskVO> getFlowWorkSubTasksRefTask(Set<Long> taskIds, int level, Integer directType, Integer appType) {
 
         //获得所有父节点task
-        List<BatchTaskShade> tasks = taskShadeService.getTaskByIds(new ArrayList<>(taskIds),appType);
+        List<ScheduleTaskShade> tasks = taskShadeService.getTaskByIds(new ArrayList<>(taskIds),appType);
         if (CollectionUtils.isEmpty(tasks)) {
             return null;
         }
 
-        List<com.dtstack.engine.api.vo.BatchTaskVO> refTaskVoList = new ArrayList<>(tasks.size());
-        for (BatchTaskShade task : tasks) {
+        List<ScheduleTaskVO> refTaskVoList = new ArrayList<>(tasks.size());
+        for (ScheduleTaskShade task : tasks) {
             refTaskVoList.add(this.getFlowWorkOffSpring(task, level, directType,appType));
         }
 
