@@ -1,7 +1,7 @@
 package com.dtstack.engine.master.job;
 
 import com.dtstack.engine.api.domain.ScheduleJob;
-import com.dtstack.engine.api.domain.BatchTaskShade;
+import com.dtstack.engine.api.domain.ScheduleTaskShade;
 
 import java.util.Map;
 
@@ -10,5 +10,5 @@ import java.util.Map;
  * @date 2019-11-05
  */
 public interface IJobStartTrigger {
-     void readyForTaskStartTrigger(Map<String, Object> actionParam, BatchTaskShade taskShade, ScheduleJob scheduleJob) throws Exception;
+     void readyForTaskStartTrigger(Map<String, Object> actionParam, ScheduleTaskShade taskShade, ScheduleJob scheduleJob) throws Exception;
 }
