@@ -8,7 +8,7 @@ import com.dtstack.engine.api.domain.ScheduleJob;
 import com.dtstack.engine.common.CustomThreadFactory;
 import com.dtstack.engine.common.exception.ErrorCode;
 import com.dtstack.engine.common.exception.RdosDefineException;
-import com.dtstack.engine.api.domain.BatchTaskShade;
+import com.dtstack.engine.api.domain.ScheduleTaskShade;
 import com.dtstack.engine.master.impl.ActionService;
 import com.dtstack.engine.master.impl.ScheduleTaskShadeService;
 import org.apache.commons.collections.CollectionUtils;
@@ -89,17 +89,17 @@ public class JobStopSender implements InitializingBean, DisposableBean, Runnable
                 .map(ScheduleJob::getTaskId)
                 .collect(Collectors.toList());
 
-        Map<Long, List<BatchTaskShade>> taskShades =
+        Map<Long, List<ScheduleTaskShade>> taskShades =
                 batchTaskShadeService.getTaskByIds(taskIds, stoppedJob.getAppType())
                         .stream()
-                        .collect(Collectors.groupingBy(BatchTaskShade::getTaskId));
+                        .collect(Collectors.groupingBy(ScheduleTaskShade::getTaskId));
 
         JSONArray jsonArray = new JSONArray();
         for (ScheduleJob job : stoppedJob.getJobs()) {
-            List<BatchTaskShade> shades = taskShades.get(job.getTaskId());
+            List<ScheduleTaskShade> shades = taskShades.get(job.getTaskId());
 
             if (CollectionUtils.isNotEmpty(shades)) {
-                BatchTaskShade batchTask = shades.get(0);
+                ScheduleTaskShade batchTask = shades.get(0);
                 JSONObject params = new JSONObject();
                 params.put("engineType", EngineType.getEngineName(batchTask.getEngineType()));
                 params.put("taskId", job.getJobId());
